requirements.txt
streamlit
rembg
Pillow
numpy